requirements.txt
flask
Fortuna
gunicorn
Werkzeug
bleach
certifi
python-dotenv
pymongo[srv]
pandas
vaderSentiment
